Fix Compile warnings
Most of the fixes here address compiler warnings.  Mostly variables that were passed but not being used, compares of unsigned to signed values.  Also added reset behavior for Ground Radio Board.

+++ e2ecommands_metro/ExternalWatchdog.h
@@ -29,8 +29,8 @@ class ExternalWatchdog
 
   private:
     int _pin;
-    int watchdog_lower_boundary {24};
-    int m_last_action_time;
+    unsigned long watchdog_lower_boundary {24};
+    unsigned long m_last_action_time;
 };
 
 #endif
